Controllo e fix/aggiunte di varie cose nei bean. In particolare ho aggiunto vari controlli alle variabili numeriche che non dovrebbero essere negative ed ho aggiunto/modificato i regex, in particolare il regex del numero di telefono. Ho notato di aver sbagliato il tipo delle variabili "totale" nelle tabelle del DB, ho cambiato da int a decimal, per cui anche nei bean è stato cambiato da int a double. Ho cambiato alcuni tipi numerici nei bean in short invece che int.

+++ src/main/java/com/tsw/studentsupps/Model/Carrello.java
@@ -5,7 +5,7 @@ import java.sql.Timestamp;
 
 public class Carrello {
     private String id;
-    private int  totale;
+    private double totale;
     private  Timestamp updated_at;
 
     public String getId() {
@@ -16,11 +16,11 @@ public class Carrello {
             this.id = id;
     }
 
-    public int getTotale() {
+    public double getTotale() {
         return totale;
     }
 
-    public void setTotale(int totale) {
+    public void setTotale(double totale) {
         this.totale = totale;
     }
